requirements.txt
accelerate==0.30.1
groq~=0.8.0
ollama~=0.2.0
optimum==1.20.0
python-dotenv~=1.0.1
telethon~=1.35.0
torch~=2.3.0
transformers~=4.41.2